requirements.txt
# Production dependencies
fastapi==0.115.6
mangum==0.19.0
boto3==1.35.80
pydantic==2.10.6
pydantic-core==2.27.2
pandas==2.2.3
gspread==6.1.4
google-auth==2.37.0
google-auth-oauthlib==1.2.1
# creditagricole-particuliers - Install manually or from layer
# Package not available on PyPI
requests==2.32.3

# Development dependencies
pytest>=8.0.0
pytest-cov>=4.1.0
flake8>=7.0.0
black>=24.1.0
httpx>=0.27.0